VLCMediaParser: make it a public API
This removes the ability to trigger parsing on a VLCMedia instance.
It needs to be queued to a parser now.

For convenience and to keep backward compatibility, the parsing part
of the VLCMedia delegation protocol and the key-value-coding is retained.

This commit also updates the related sample code project.
