--- Headers/Internal/VLCMediaParser.h
@@ -1,38 +1,0 @@
-/*****************************************************************************
- * VLCMediaParser.h
- *****************************************************************************
- * Copyright (C) 2024 VLC authors and VideoLAN
- *
- * Authors: Felix Paul Kühne <fkuehne # videolan.org
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation; either version 2.1 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston MA 02110-1301, USA.
- *****************************************************************************/
-
-#import <Foundation/Foundation.h>
-#import "VLCMedia.h"
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface VLCMediaParser : NSObject
-
-+ (instancetype)sharedParser;
-
-- (instancetype)initWithLibrary:(VLCLibrary *)library;
-
-- (int)queueMedia:(VLCMedia *)media withDescriptor:(libvlc_media_t *)md options:(int)options timeout:(int)timeout;
-
-@end
-
-NS_ASSUME_NONNULL_END
